Moved delete logic
