+++ src/components/ThumbnailControls/ThumbnailControls.js
@@ -1,27 +1,30 @@
 import React from 'react';
-import { useSelector } from 'react-redux';
+import { useDispatch, useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
+import { useTranslation } from 'react-i18next';
 
 import Button from 'components/Button';
 import core from 'core';
 import selectors from 'selectors';
+import actions from 'actions';
 
 import './ThumbnailControls.scss';
 
 const propTypes = {
   index: PropTypes.number.isRequired,
-  handleDelete: PropTypes.func.isRequired,
 };
 
 const dataElementName = 'thumbnailControl';
 
 const ThumbnailControls = ({
   index,
-  handleDelete,
 }) => {
   const [isElementDisabled] = useSelector(state => [
     selectors.isElementDisabled(state, dataElementName),
   ]);
+
+  const [t] = useTranslation();
+  const dispatch = useDispatch();
 
   const rotateClockwise = () => {
     core.rotatePages([index + 1], window.CoreControls.PageRotation.e_90);
@@ -29,6 +32,32 @@ const ThumbnailControls = ({
 
   const rotateCounterClockwise = () => {
     core.rotatePages([index + 1], window.CoreControls.PageRotation.e_270);
+  };
+
+  const handleDelete = () => {
+    let message = t('option.thumbnailPanel.deleteWarningMessage');
+    const title = t('option.thumbnailPanel.deleteWarningTitle');
+    const confirmBtnText = t('action.ok');
+
+    let warning = {
+      message,
+      title,
+      confirmBtnText,
+      onConfirm: () => core.removePages([index + 1]),
+    };
+
+    if (core.getDocumentViewer().getPageCount() === 1) {
+      message = t('option.thumbnailPanel.deleteLastPageError');
+
+      warning = {
+        message,
+        title,
+        confirmBtnText,
+        onConfirm: () => Promise.resolve(),
+      };
+    }
+
+    dispatch(actions.showWarningMessage(warning));
   };
 
   return isElementDisabled ? null : (
